organized into folders and files
